v3.32 (May 11, 2020)
* SECURITY: Fixes CVE-2020-1763 https://libreswan.org/security/CVE-2020-1763
* IKEv2: Support non-narrowed child rekey for narrowing (regression in 3.31)
* FIPS: ECDSA keys were mistakenly rejected as "too weak" [Paul]
* FIPS: Minimum RSA key size is 2048, not 3072 [Paul]
* FIPS: Use NSS to check FIPS mode instead of manually checking fips=1 [Paul]
* IKEv2: Do not use fragments if not appropriate (regression from v3.30) [Paul]
* IKEv1: Add NSS KDF support for the Quick Mode KDF [Andrew/Paul]
* libipsecconf: support old-style ",," to mean "\," in specifying id [Paul]
* libipsecconf: left/rightinterface-ip= are not kt_obsolete [Paul]
* whack: Add missing ecdsa/sha2 and compat rsa policy options to whack [Paul]
* Fix left=%iface syntax due to string length miscalculation [Antony]
* X509: don't try to match up ID on SAN when ID type is ID_DER_ASN1_DN [Paul]
* packaging: debian fixes [Antony]
* building: USE_NSS_KDF=true now uses NSS for all KDF functions
            Using this option, libreswan no longer needs FIPS certification

+++ mk/docker-targets.mk
@@ -9,7 +9,7 @@
 # The variables above could be set from command line.
 #
 
-DOCKER_CMD ?= sudo docker
+DOCKER_CMD ?= sudo podman
 D ?= testing/docker
 
 DI_T ?= swanbase 	#docker image tag
@@ -17,59 +17,98 @@ DI_T ?= swanbase 	#docker image tag
 W1 = $(firstword $(subst -, ,$1))
 W2 = $(or $(word 2, $(subst -, ,$1)), $(value 2))
 W3 = $(or $(word 3, $(subst -, ,$1)), $(value 2))
+W4 = $(or $(word 4, $(subst -, ,$1)), $(value 2))
 
 FIRST_TARGET ?=$@	# keep track of original target
 DISTRO ?= fedora	# default distro
-DISTRO_REL ?= 28 	# default release
-EXCLUDE_RPM_ARCH ?= --excludepkgs='*.i686'
+DISTRO_REL ?= 30 	# default release
+
+D_USE_UNBOUND_EVENT_H_COPY ?= true
+D_USE_DNSSEC ?= false
+D_USE_NSS_IPSEC_PROFILE ?= flase
+D_USE_GLIBC_KERN_FLIP_HEADERS ?= true
+D_USE_NSS_AVA_COPY ?= true
+
+DOCKERFILE ?= $(D)/dockerfile
+
+SUDO_CMD ?= sudo
+PKG_CMD = $(shell test -f /usr/bin/dnf && echo /usr/bin/dnf || (test -f /usr/bin/yum && echo /usr/bin/yum || echo "no yum or dnf found" && exit 1))
+PKG_BUILDDEP = $(shell test -f /usr/bin/dnf && echo "/usr/bin/dnf builddep" || echo /usr/bin/yum-builddep )
+REPO_POWERTOOLS = $(shell grep -qE '^ID="centos"' /etc/os-release && echo "--enablerepo=PowerTools" || echo "" )
+PKG_INSTALL = $(SUDO_CMD) $(PKG_CMD) $(REPO_POWERTOOLS) install -y
+PKG_UPGRADE = $(SUDO_CMD) $(PKG_CMD) $(REPO_POWERTOOLS) upgrade -y
+PKG_DEBUGINFO_INSTALL = $(SUDO_CMD) $(PKG_CMD) $(REPO_POWERTOOLS) debuginfo-install -y
+PKG_DEBUGINFO_UPGRADE = $(SUDO_CMD) $(PKG_CMD) $(REPO_POWERTOOLS) upgrade --enablerepo=*-debuginfo "*-debuginfo" -y
 
 # end of configurable variables
 
 DI = $(DISTRO)-$(DISTRO_REL)
-DOCKERFILE ?= $(D)/dockerfile
 DOCKERFILE_PKG = $(D)/Dockerfile-$(DISTRO)-min-packages
 TWEAKS=
+LOCAL_MAKE_FLAGS=
+MAKE_BASE = base
+MAKE_INSTLL_BASE = install-base
+
+ifdef TRAVIS_ENABLED
+BRANCH = $(shell test -d .git -o -f .git && (git rev-parse --abbrev-ref HEAD || echo ''))
+TRAVIS_BANCH ?= $(call W1, $(BRANCH),'')
+endif
+ifeq ($(TRAVIS_BANCH), travis)
+	DISTRO =  $(call W2, $(BRANCH),fedora)
+	DISTRO_REL = $(call W3, $(BRANCH),27)
+endif
 
 #
 # Distribution specific tweaks
 #
 ifeq ($(DISTRO), ubuntu)
+	MAKE_BASE = deb
+	MAKE_INSTLL_BASE = deb-install
 	DOCKERFILE_PKG=$(D)/Dockerfile-debian-min-packages
-	TWEAKS = dockerfile-ubuntu-cmd
-	ifeq ($(DISTRO_REL), xenial)
-		TWEAKS = flip-glibc-kern-headers use_unbound_event_h_copy enable-nss_ava_copy disable-nss_ipsec_profile
-	endif
-	ifeq ($(DISTRO_REL), cosmic)
-		TWEAKS = disable-nss_ipsec_profile
-	endif
-	ifeq ($(DISTRO_REL), bionic)
-		TWEAKS = disable-nss_ipsec_profile
-	endif
 endif
 
 ifeq ($(DISTRO), debian)
 	DOCKERFILE_PKG=$(D)/Dockerfile-debian-min-packages
 	TWEAKS = dockerfile-debian-cmd
+	MAKE_BASE = deb
+	MAKE_INSTLL_BASE = deb-install
 endif
 
 ifeq ($(DISTRO), centos)
-	ifeq ($(DISTRO_REL), 7)
-		TWEAKS = werror-no-missing-field-initializers
-	endif
+	MAKE_BASE = base
+	MAKE_INSTLL_BASE = install-base
 	ifeq ($(DISTRO_REL), 6)
-		TWEAKS = werror-no-missing-field-initializers disable-dsnssec disable-nss_ipsec_profile
+		DOCKERFILE_PKG = $(D)/Dockerfile-$(DISTRO)6-min-packages
+		LOCAL_MAKE_FLAGS += USE_DNSSEC=$(D_USE_DNSSEC)
+		LOCAL_MAKE_FLAGS += USE_NSS_IPSEC_PROFILE=$(D_USE_NSS_IPSEC_PROFILE)
+		LOCAL_MAKE_FLAGS += USE_XFRM_INTERFACE_IFLA_HEADER=true
 	endif
+
+	ifeq ($(DISTRO_REL), 7)
+		LOCAL_MAKE_FLAGS += USE_XFRM_INTERFACE_IFLA_HEADER=true
+	endif
+
+	ifeq ($(DISTRO_REL), 8)
+		# CentOS 8 Fedora 28 based so it should be able to handle basic build
+		DOCKERFILE_PKG = $(D)/Dockerfile-fedora-min-packages
+	endif
+
 endif
 
-#ifeq ($(DISTRO), fedora)
-#endif
+ifeq ($(DISTRO), fedora)
+	ifeq ($(DISTRO_REL), rawhide)
+		TWEAKS += rawhide-remove-dnf-update
+	endif
 
-BRANCH = $(shell test -d .git && test -f /usr/bin/git -o -f /usr/local/bin/git && git rev-parse --abbrev-ref HEAD)
-TRAVIS_BANCH = $(call W1, $(BRANCH),'')
-ifeq ($(TRAVIS_BANCH), travis)
-	DISTRO =  $(call W2, $(BRANCH),fedora)
-	DISTRO_REL = $(call W3, $(BRANCH),27)
+	MAKE_BASE = base
+	MAKE_INSTLL_BASE = install-base
+	LOCAL_MAKE_FLAGS =
 endif
+
+.PHONY: rawhide-remove-dnf-update
+rawhide-remove-dnf-update:
+	# on rawhide RUN dnf -y update could be a bad idea
+	$(shell sed -i '/RUN dnf -y update/d' testing/docker/dockerfile)
 
 .PHONY: dockerfile-remove-libreswan-spec
 dockerfile-remove-libreswan-spec:
@@ -83,50 +122,38 @@ dockerfile-debian-cmd:
 dockerfile-ubuntu-cmd:
 	$(shell sed -i 's#CMD.*#CMD ["/sbin/init"]#' testing/docker/dockerfile)
 
-.PHONY: disable-dsnssec
-disable-dsnssec:
-	$(shell (grep "^USE_DNSSEC" Makefile.inc.local || echo "USE_DNSSEC ?= false" >> Makefile.inc.local))
-
-.PHONY: disable-nss_ipsec_profile
-disable-nss_ipsec_profile:
-	$(shell (grep "^USE_NSS_IPSEC_PROFILE" Makefile.inc.local || echo "USE_NSS_IPSEC_PROFILE ?= false" >>Makefile.inc.local))
-
-.PHONY: enable-nss_ava_copy
-enable-nss_ava_copy:
-	$(shell (grep "^USE_NSS_AVA_COPY" Makefile.inc.local || echo "USE_NSS_AVA_COPY ?= true" >> Makefile.inc.local))
-
-
-.PHONY: werror-no-missing-field-initializers
-werror-no-missing-field-initializers:
-	 $(shell (grep "^WERROR_CFLAGS" Makefile.inc.local || echo "WERROR_CFLAGS ?= -Werror -Wno-missing-field-initializers" >> Makefile.inc.local))
-
-.PHONY: disable-seccomp
-disable-seccomp:
-	$(shell (grep "^USE_SECCOMP" Makefile.inc.local || echo "USE_SECCOMP ?= false" >> Makefile.inc.local))
-
-.PHONY: flip-glibc-kern-headers
-flip-glibc-kern-headers:
-	$(shell (grep "^USE_GLIBC_KERN_FLIP_HEADERS" Makefile.inc.local || echo "USE_GLIBC_KERN_FLIP_HEADERS ?= true" >> Makefile.inc.local))
-
 .PHONY: use_unbound_event_h_copy
 use_unbound_event_h_copy:
-	$(shell (grep "^USE_UNBOUND_EVENT_H_COPY" Makefile.inc.local || echo "USE_UNBOUND_EVENT_H_COPY ?= true" >> Makefile.inc.local))
 
 #
 # end  of Distribution tweaks
 #
 
-.PHONY: install-rpm-dep
-RUN_RPMS = $$(dnf deplist $(EXCLUDE_RPM_ARCH) libreswan | awk '/provider:/ {print $$2}' | sort -u)
-install-rpm-dep:
-	$(if $(KVM_PACKAGES), $(KVM_PACKAGE_INSTALL) $(KVM_PACKAGES))
-	$(if $(KVM_PACKAGES), $(KVM_PACKAGE_UPGRADE) $(KVM_PACKAGES))
-	dnf builddep -y libreswan
-	dnf install -y \@development-tools
-	dnf install -y --skip-broken $(RUN_RPMS)
+.PHONY: install-testing-rpm-dep
+install-testing-rpm-dep: install-rpm-build-dep install-rpm-run-dep
+	$(if $(KVM_INSTALL_PACKAGES), $(PKG_INSTALL) $(KVM_INSTALL_PACKAGES))
+	$(if $(KVM_UPGRADE_PACKAGES), $(PKG_UPGRADE) $(KVM_UPGRADE_PACKAGES))
+	$(if $(KVM_DEBUGINFO_INSTALL), $(if $(KVM_DEBUGINFO), \
+                $(PKG_DEBUGINFO_INSTALL) $(KVM_DEBUGINFO)))
+	$(if $(KVM_DEBUGINFO_INSTALL), $(if $(KVM_DEBUGINFO), \
+		$(PKG_DEBUGINFO_UPGRADE)))
+
+.PHONY: install-rpm-run-dep
+RUN_RPMS = $$($(PKG_CMD) deplist --arch $$(uname -m) --forcearch $$(uname -m) libreswan | awk '/provider:/ {print $$2}' | sort -u)
+install-rpm-run-dep:
+	 $(PKG_INSTALL) --skip-broken $(RUN_RPMS)
+
+.PHONY: install-rpm-build-dep
+install-rpm-build-dep:
+	$(SUDO_CMD) $(PKG_CMD) groupinstall $(POWER_TOOLS) -y 'Development Tools'
+	$(SUDO_CMD) $(PKG_BUILDDEP) $(REPO_POWERTOOLS) -y libreswan
 
 .PHONY: install-deb-dep
-RUN_DEBS = $$(test -f /usr/bin/apt-cache1 && apt-cache depends libreswan | awk '/Depends:/{print $$2}' | grep -v "<" | sort -u)
+# RUN_DEBS_OLD ?= $$(grep -qE 'jessie|xenial' /etc/os-release && echo "host iptables")
+# hard codde these two packages it fail on xenial and old ones.
+# on buster host is virtual package
+RUN_DEBS_OLD ?= bind9-host iptables
+RUN_DEBS ?= $$(test -f /usr/bin/apt-cache && apt-cache depends libreswan | awk '/Depends:/{print $$2}' | grep -v "<" | sort -u)
 install-deb-dep:
 	apt-get update
 	# development dependencies
@@ -136,13 +163,17 @@ install-deb-dep:
 	cp -r packaging/debian/control libreswan-control
 	mk-build-deps --install --tool "apt-get -o Dpkg::Options::="--force-confold" -o Debug::pkgProblemResolver=yes -y --no-install-recommends" libreswan-control
 	# install libreswan runtime dependencies
-	apt-get install -y $(RUN_DEBS)
+	apt-get install -y $(RUN_DEBS) $(RUN_DEBS_OLD)
 	# give another kick
-	apt --fix-broken install -y 
+	apt-get --fix-broken install -y
 
 .PHONY: travis-docker-image
 travis-docker-image:
 	$(MAKE) DISTRO=$(DISTRO) DISTRO_REL=$(DISTRO_REL) docker-image
+
+.PHONY: travis-docker-base
+travis-docker-base:
+	$(MAKE) $(MAKE_BASE)
 
 define debian_exp_repo
 	if [ $(1) == "experimental" ] ; then \
@@ -167,6 +198,8 @@ travis-ubuntu-xenial: ubuntu-xenial-packages
 
 .PHONY: docker-build
 docker-build: dockerfile
+	# --volume is only in podman
+	# $(DOCKER_CMD) build -t $(DI_T) --volume /home/build/libreswan:/home/build/libreswan -f $(DOCKERFILE) .
 	$(DOCKER_CMD) build -t $(DI_T) -f $(DOCKERFILE) .
 
 .PHONY: docker-ssh-image
@@ -182,17 +215,77 @@ docker-min-image: dockerfile $(TWEAKS) docker-build
 docker-image: dockerfile $(TWEAKS) docker-ssh-image docker-build
 	echo "done docker image tag $(DI_T) from $(DISTRO)-$(DISTRO_REL) with ssh"
 
-.PHONY: docker-start
-docker-start:
-	$(DOCKER_CMD) run -h $(DI_T) --privileged --name $(DI_T) \
-	-v /home/build:/home/build \
-	-v /sys/fs/cgroup:/sys/fs/cgroup:ro -d $(DI_T)
+
+# NEW tragets to get docker handling 201906
+.PHONY: docker-instance-name
+docker-instance-name:
+	echo $(DI_T)
+
+.PHONY: travis-docker-make
+travis-docker-make:
+	$(DOCKER_CMD) exec -ti $(DI_T) /bin/bash -c "cd /home/build/libreswan && $(MAKE) DISTRO=$(DISTRO) DISTRO_REL=$(DISTRO_REL) make-base"
+
+.PHONY: travis-docker-make-install
+travis-docker-make-install:
+	$(DOCKER_CMD) exec -ti $(DI_T) /bin/bash -c "cd /home/build/libreswan && $(MAKE) DISTRO=$(DISTRO) DISTRO_REL=$(DISTRO_REL) make-install"
 
 .PHONY: docker-exec
 docker-exec:
-	$(DOCKER_CMD) exec -ti $(DI_T) /bin/bash
+	$(DOCKER_CMD) exec -ti $(DI_T) /bin/bash -c "cd /home/build/libreswan && $(MAKE) $(1)"
 
 .PHONY: docker-stop
 docker-stop:
-	$(DOCKER_CMD) stop $(DI_T)
-	$(DOCKER_CMD) rm $(DI_T)
+	$(DOCKER_CMD) stop $(DI_T) && $(DOCKER_CMD) rm $(DI_T) || echo "nothing to stop $(DI_T)"
+
+.PHONY: docker-shell
+docker-shell:
+	$(DOCKER_CMD) exec -ti $(DI_T) /bin/bash
+
+.PHONY: make-base
+make-base:
+	$(LOCAL_MAKE_FLAGS) $(MAKE) $(MAKE_BASE)
+
+.PHONY: make-install
+make-install:
+	$(LOCAL_MAKE_FLAGS) $(MAKE) $(MAKE_INSTLL_BASE)
+
+.PHONY: deb-install
+deb-install: install-deb-dep
+	dpkg -i  ../*.deb || apt-get --fix-broken install -y
+
+.PHONY: docker-make-install
+docker-make-install: docker-stop
+	$(DOCKER_CMD) run --privileged --net=none --name $(DI_T) \
+		-v $(PWD):/home/build/libreswan \
+		-v /sys/fs/cgroup:/sys/fs/cgroup:ro \
+		-ti $(DI_T) /bin/bash -c "cd /home/build/libreswan && $(MAKE) $(MAKE_INSTLL_BASE)"
+
+.PHONY: docker-make-base
+docker-make-base: docker-stop
+	$(DOCKER_CMD) run --privileged --net=none --name $(DI_T) \
+		-v $(PWD):/home/build/libreswan \
+		-v /sys/fs/cgroup:/sys/fs/cgroup:ro \
+		-ti $(DI_T) /bin/bash -c "cd /home/build/libreswan && \
+		$(LOCAL_MAKE_FLAGS) $(MAKE) $(MAKE_BASE)"
+
+.PHONY: travis-docker-start
+travis-docker-start:
+	$(DOCKER_CMD) run -h $(DI_T) --privileged  --name $(DI_T) \
+		-v $(PWD):/home/build/libreswan/ \
+		-v /sys/fs/cgroup:/sys/fs/cgroup:ro -d $(DI_T)
+
+.PHONY: nsrunclean
+nsrunclean:
+	sudo rm -fr $(abs_top_srcdir)/testing/pluto/*/OUTPUT $(abs_top_srcdir)/testing/pluto/*/NS
+
+NSURNDIRS = $(shell mount | grep "^nsfs" | cut -d  " " -f 3)
+.PHONY: nsrun
+nsrun: nsrunclean
+	$(if $(NSURNDIRS), $(shell sudo umount $(NSURNDIRS)), $(echo "no nsfs"))
+	$(abs_top_srcdir)/testing/utils/nsrun  --ns --shutdown --log-level debug --verbos 2 --testrun
+
+.PHONY: nsinstall
+nsinstall:
+	$(MAKE) clean
+	$(MAKE) INITSYSTEM=docker DOCKER_PLUTONOFORK= base
+	sudo $(MAKE) INITSYSTEM=docker DOCKER_PLUTONOFORK= install-base
